ai/Readee_AI_System/requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-multipart==0.0.6
python-dotenv==1.0.1
aiofiles==23.2.1

# Transformers & LLM
transformers==4.45.2
accelerate==0.34.2
bitsandbytes==0.43.3
sentencepiece==0.2.0
peft==0.12.0
huggingface-hub>=0.20.0

# PDF / DOCX / Image processing
PyMuPDF==1.23.8
python-docx==1.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Data & ML utils
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
matplotlib==3.9.2

# Monitoring / utilities
tensorboard==2.17.1
tensorboardX==2.6.2.2
psutil==6.0.0
tqdm==4.66.5

# Safety & typing
pydantic==2.8.2
typing-extensions>=4.11.0
pyyaml>=6.0